requirements.txt
fastapi~=0.115.12
uvicorn
pydantic~=2.11.4
passlib~=1.7.4
pytest~=8.3.5
requests~=2.32.3
PyJWT~=2.10.1
starlette~=0.46.2
pika
Pillow
bcrypt~=3.2.2
psycopg2-binary
sqlalchemy
redis[hiredis]
pydantic_settings
prometheus_fastapi_instrumentator
prometheus-client
sqlalchemy_utils